% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_sfplsr.R
\name{predict_sfplsr}
\alias{predict_sfplsr}
\title{Predict Method for Spatial FPLS Regression}
\usage{
predict_sfplsr(object, xnew, wnew)
}
\arguments{
\item{object}{An object obtained from the \code{splsr} function.}

\item{xnew}{A numeric matrix of new functional predictors (new observations).}

\item{wnew}{A square spatial weights matrix for the new observations or test validation block.}
}
\value{
A numeric vector of predicted values for the scalar response.
}
\description{
Predict Method for Spatial FPLS Regression
}
