% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfplsr.R
\name{sfplsr}
\alias{sfplsr}
\title{Spatial Functional Partial Least Squares Regression}
\usage{
sfplsr(
  y,
  x,
  h,
  nbasis = NULL,
  gpx = NULL,
  wei_mat,
  rho_grid = seq(-0.95, 0.95, by = 0.01),
  criterion = c("train_mse", "bic", "aic"),
  scv = FALSE,
  block_ids = NULL,
  verbose = TRUE
)
}
\arguments{
\item{y}{A numeric vector of the scalar response variable.}

\item{x}{A numeric matrix of functional explanatory variable.}

\item{h}{An integer or vector specifying the number of latent components.}

\item{nbasis}{An integer or vector specifying the number of basis functions.}

\item{gpx}{A numeric vector defining the evaluation grid points for \code{x}.}

\item{wei_mat}{A square spatial weights matrix.}

\item{rho_grid}{A numeric vector defining the search grid for spatial parameter \eqn{\rho}.}

\item{criterion}{Optimization metric: \code{"train_mse"}, \code{"bic"}, or \code{"aic"}.}

\item{scv}{Logical; if \code{TRUE}, performs Spatial Cross-Validation for \code{nbasis} and \code{h}.}

\item{block_ids}{A vector indicating the spatial block assignment for each observation (required if \code{scv = TRUE}).}

\item{verbose}{Logical; if \code{TRUE}, prints progress and tuning results.}
}
\description{
Spatial Functional Partial Least Squares Regression
}
