% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_data_generation.R
\name{spatial_data_generation}
\alias{spatial_data_generation}
\title{Generate Spatial Functional Data}
\usage{
spatial_data_generation(n, j, rho, sig.e)
}
\arguments{
\item{n}{An integer specifying the sample size (number of spatial locations).}

\item{j}{An integer specifying the number of grid points for the functional predictor.}

\item{rho}{A numeric value representing the spatial autoregressive parameter.}

\item{sig.e}{A numeric value specifying the standard deviation of the error terms.}
}
\value{
A list containing the generated spatial functional dataset components.
}
\description{
Generate Spatial Functional Data
}
