% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UK_weather_data}
\alias{UK_weather_data}
\title{UK Weather Data for Spatial Functional Partial Least Squares Regression}
\format{
A named list containing the weather station variables.
}
\source{
NASA POWER Dataset
}
\usage{
UK_weather_data
}
\description{
A dataset containing solar radiation, temperature, and station info
from UK weather stations used in the SFPLS framework.
}
\keyword{datasets}
